# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/entrar": {
      "post": {
        "operationId": "IdentidadeController_entrar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntrarDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Sessão aberta; os cookies vão na resposta."
          },
          "401": {
            "description": "E-mail ou senha incorretos."
          }
        },
        "summary": "Abre uma sessão",
        "tags": [
          "identidade"
        ]
      }
    },
...
    "/api/v1/auth/senha": {
      "post": {
        "operationId": "IdentidadeController_trocarSenha",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrocarSenhaDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Senha trocada; é preciso entrar de novo."
          },
          "401": {
            "description": "Senha atual incorreta."
          }
        },
        "summary": "Troca a própria senha e derruba as outras sessões",
        "tags": [
          "identidade"
        ]
      }
    },
    "/api/v1/auth/usuarios": {
      "get": {
        "operationId": "IdentidadeController_listarUsuarios",
        "parameters": [
          {
            "name": "papel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ADMIN",
                "VETERINARIO",
                "FARMACIA",
                "CLINICA"
              ]
            }
          },
          {
            "name": "busca",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 160,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListaDeUsuariosDto"
                }
              }
            }
          },
          "403": {
            "description": "Só administrador lista usuários."
          }
        },
        "summary": "Lista os usuários",
        "tags": [
          "identidade"
        ]
      },
      "post": {
        "operationId": "IdentidadeController_criarUsuario",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarUsuarioDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EuDto"
                }
              }
            }
          },
          "403": {
            "description": "Só administrador cria usuário."
          }
        },
        "summary": "Cria um usuário",
        "tags": [
          "identidade"
        ]
      }
    },
...
    "/api/v1/receituario/receitas": {
      "post": {
        "operationId": "ReceitaController_criar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalvarReceitaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceitaDto"
                }
              }
            }
          }
        },
        "summary": "Abre um rascunho de receita",
        "tags": [
          "receituário"
        ]
      },
      "get": {
        "operationId": "ReceitaController_listar",
        "parameters": [
          {
            "name": "pacienteId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "estado",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "RASCUNHO",
                "EMITIDA",
                "CANCELADA"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListaDeReceitasDto"
                }
              }
            }
          }
        },
        "summary": "Lista receitas",
        "tags": [
          "receituário"
        ]
      }
    },
    "/api/v1/receituario/receitas/{id}": {
      "get": {
        "operationId": "ReceitaController_achar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceitaDto"
                }
              }
            }
          },
          "404": {
            "description": "Não existe, ou não é visível para quem perguntou."
          }
        },
        "summary": "A receita inteira, com fórmulas, preço e avisos",
        "tags": [
          "receituário"
        ]
      },
      "put": {
        "operationId": "ReceitaController_salvar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalvarReceitaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceitaDto"
                }
              }
            }
          },
          "400": {
            "description": "Receita emitida não se altera."
          }
        },
        "summary": "Regrava o rascunho",
        "tags": [
          "receituário"
        ]
      }
    },
    "/api/v1/receituario/receitas/{id}/emitir": {
      "post": {
        "operationId": "ReceitaController_emitir",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceitaDto"
                }
              }
            }
          },
          "400": {
            "description": "Falta CRMV, falta peso do paciente, há impedimento na fórmula ou lista sem prazo."
          }
        },
        "summary": "Emite a receita: numera, congela e passa a valer",
        "tags": [
          "receituário"
        ]
      }
    },
    "/api/v1/receituario/receitas/{id}/cancelar": {
      "post": {
        "operationId": "ReceitaController_cancelar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelarReceitaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceitaDto"
                }
              }
            }
          }
        },
        "summary": "Cancela a receita, com motivo",
        "tags": [
          "receituário"
        ]
      }
    },
...
  },
  "info": {
    "title": "PharmoPet API",
    "description": "Prescrição e manipulação veterinária",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "EntrarDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "senha": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "senha"
        ]
      },
      "EuDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "papel": {
            "type": "string",
            "enum": [
              "ADMIN",
              "VETERINARIO",
              "FARMACIA",
              "CLINICA"
            ]
          },
          "crmv": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "nome",
          "email",
          "papel",
          "crmv"
        ]
      },
      "TrocarSenhaDto": {
        "type": "object",
        "properties": {
          "senhaAtual": {
            "type": "string",
            "minLength": 1
          },
          "senhaNova": {
            "type": "string",
            "minLength": 12,
            "maxLength": 200
          }
        },
        "required": [
          "senhaAtual",
          "senhaNova"
        ]
      },
      "ListaDeUsuariosDto": {
        "type": "object",
        "properties": {
          "usuarios": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "nome": {
                  "type": "string"
                },
                "email": {
                  "type": "string",
                  "format": "email",
                  "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                },
                "papel": {
                  "type": "string",
                  "enum": [
                    "ADMIN",
                    "VETERINARIO",
                    "FARMACIA",
                    "CLINICA"
                  ]
                },
                "crmv": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "bloqueado": {
                  "type": "boolean"
                },
                "criadoEm": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                }
              },
              "required": [
                "id",
                "nome",
                "email",
                "papel",
                "crmv",
                "bloqueado",
                "criadoEm"
              ]
            }
          }
        },
        "required": [
          "usuarios"
        ]
      },
      "CriarUsuarioDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "nome": {
            "type": "string",
            "minLength": 2,
            "maxLength": 160
          },
          "papel": {
            "type": "string",
            "enum": [
              "ADMIN",
              "VETERINARIO",
              "FARMACIA",
              "CLINICA"
            ]
          },
          "senha": {
            "type": "string",
            "minLength": 12,
            "maxLength": 200
          },
          "crmv": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "email",
          "nome",
          "papel",
          "senha"
        ]
      },
...
      "SalvarReceitaDto": {
        "type": "object",
        "properties": {
          "pacienteId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "clinicaId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              {
                "type": "null"
              }
            ]
          },
          "observacoes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ]
          },
          "formulacoes": {
            "minItems": 1,
            "maxItems": 10,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "formaId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "frequenciaHoras": {
                  "anyOf": [
                    {
                      "type": "number",
                      "const": 24
                    },
                    {
                      "type": "number",
                      "const": 12
                    },
                    {
                      "type": "number",
                      "const": 8
                    },
                    {
                      "type": "number",
                      "const": 6
                    }
                  ]
                },
                "dias": {
                  "type": "integer",
                  "exclusiveMinimum": 0,
                  "maximum": 365
                },
                "quantidade": {
                  "type": "integer",
                  "exclusiveMinimum": 0,
                  "maximum": 1000
                },
                "orientacao": {
                  "anyOf": [
                    {
                      "type": "string",
                      "maxLength": 500
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "itens": {
                  "minItems": 1,
                  "maxItems": 20,
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "insumoId": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "doseMg": {
                        "type": "number",
                        "exclusiveMinimum": 0,
                        "maximum": 100000
                      }
                    },
                    "required": [
                      "insumoId",
                      "doseMg"
                    ]
                  }
                }
              },
              "required": [
                "formaId",
                "frequenciaHoras",
                "dias",
                "itens"
              ]
            }
          }
        },
        "required": [
          "pacienteId",
          "formulacoes"
        ]
      },
...
      "ListaDeReceitasDto": {
        "type": "object",
        "properties": {
          "receitas": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "numero": {
                  "anyOf": [
                    {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "estado": {
                  "type": "string",
                  "enum": [
                    "RASCUNHO",
                    "EMITIDA",
                    "CANCELADA"
                  ]
                },
                "situacao": {
                  "type": "string",
                  "enum": [
                    "rascunho",
                    "valida",
                    "vencida",
                    "cancelada"
                  ]
                },
                "pacienteNome": {
                  "type": "string"
                },
                "tutorNome": {
                  "type": "string"
                },
                "veterinarioNome": {
                  "type": "string"
                },
                "emitidaEm": {
                  "anyOf": [
                    {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "validaAte": {
                  "anyOf": [
                    {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "criadaEm": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                }
              },
              "required": [
                "id",
                "numero",
                "estado",
                "situacao",
                "pacienteNome",
                "tutorNome",
                "veterinarioNome",
                "emitidaEm",
                "validaAte",
                "criadaEm"
              ]
            }
          }
        },
        "required": [
          "receitas"
        ]
      },
...
    }
  }
}